fix(studio): log resize/rotate commit failures, move anchor accumulator to resize-local
