+++ packages/studio/src/components/editor/domEditResizeLocal.ts
@@ -111,3 +111,15 @@ export function resolveRotatedResizeCursor(handle: ResizeHandle, rotationDeg: nu
   const bucket = Math.round(normalized / 45) % 8;
   return CURSORS_8[bucket]!;
 }
+
+/** Per-frame anchored-resize center accumulator: ADD the residual center correction
+ *  (fixedStart − fixedNow) onto the previous anchor so the pin CONVERGES instead of
+ *  oscillating (fa4f39168). Pure; exported for the release-shift characterization tests. */
+export function computeNextResizeAnchor(
+  prev: { dx: number; dy: number } | undefined,
+  fixedStart: { x: number; y: number },
+  fixedNow: { x: number; y: number },
+): { dx: number; dy: number } {
+  const base = prev ?? { dx: 0, dy: 0 };
+  return { dx: base.dx + (fixedStart.x - fixedNow.x), dy: base.dy + (fixedStart.y - fixedNow.y) };
+}
